+++ src/nrniv/nrncore_write/data/cell_group.h
@@ -31,8 +31,8 @@ class CellGroup {
     int* ml_vdata_offset;
     // following three are parallel arrays
     std::vector<PreSyn*> output_ps;  // n_presyn of these, real are first, tml order for acell.
-    std::vector<int> output_gid;     // n_presyn of these, -(type + 1000*index) if no gid
-    std::vector<int> output_vindex;  // n_presyn of these. >=0 if associated with voltage,
+    std::vector<long long> output_gid;     // n_presyn of these, -(type + 1000*index) if no gid
+    std::vector<long long> output_vindex;  // n_presyn of these. >=0 if associated with voltage,
                                      // -(type + 1000*index) for acell.
     int n_netcon;  // all that have targets associated with this threads Point_process.
     std::vector<NetCon*> netcons;
